requirements.txt
asgiref==3.6.0
autopep8==2.0.2
crispy-tailwind==0.5.0
Django==4.1.7
django-crispy-forms==2.0
django-resized==1.0.2
django-taggit==3.1.0
Pillow==9.4.0
pycodestyle==2.10.0
python-decouple==3.8
sqlparse==0.4.3
tomli==2.0.1
